refactor (utils): Rename BookingColors to Colors
